requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.0.0
click==8.3.0
colorama==0.4.6
colorlog==6.8.2
exceptiongroup==1.3.0
fastapi==0.115.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-multipart==0.0.20
psycopg2-binary==2.9.11
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.1
watchfiles==1.1.1
websockets==15.0.1
dotenv
